multithreaded random position

+++ src/random_position.hpp
@@ -4,4 +4,4 @@
 #include "cube.hpp"
 
 
-std::vector<State> RandomPositions (const size_t& num_elements, size_t seed_offset);
+std::vector<State> RandomPositions (const int& num_elements, int seed_offset);
